Remove old MH code.
